# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NovelTea API
  description: |
    Self-hosted, offline-first long-form writing app. Every client keeps a full local replica; this API is a synchronisation point and a compile engine, not where the work lives.

    A few things about the shape of the responses that are easy to get wrong from the schema alone:

    - **A resource the caller may not see returns `404`, never `403`** — a 403 would confirm it exists. Unauthenticated is `401`, so a client knows whether refreshing its token would help.
    - **`POST /projects/{id}/sync` answers `200` even when some changes conflict.** A conflict is an ordinary outcome, reported per change in the response body (`applied` / `conflicts`), not an HTTP error. A `document` conflict never overwrites or merges — the server keeps its version and the client's is preserved as a sibling "conflict copy" (`conflictCopyId`); see `ConflictReason` for why an individual change was rejected.
    - **`GET /projects/{id}/sync` can answer `resyncRequired: true`.** The client's cursor points into history the server can no longer explain (purged, or the project was restored from backup) and must discard it, rebuild from `GET /binder` plus documents, and resume at the returned `latestId`.
...
    - **Compiling is asynchronous.** `POST /projects/{id}/compile` returns a job id immediately; poll `GET /compile-jobs/{id}` until its status is terminal, then `GET /compile-jobs/{id}/download`.
...
  version: v1
servers:
- url: /
...
security:
- bearerAuth: []
...
paths:
...
  /api/v1/projects/{projectId}/compile:
    post:
      tags:
      - Compile Jobs
      summary: Queue a compile job
      description: |-
        Returns immediately with a job id; a separate worker process does the actual rendering. Poll GET /compile-jobs/{jobId} for status, then GET /compile-jobs/{jobId}/download once it reports a terminal status.

        Answers 501 (unavailable_in_this_edition) for a format or destination this build does not ship — this is Core, which ships only md/html/txt and the `download`/`server` destinations; `cloud` and every other export format are commercial. Identical pending requests are deduplicated: resubmitting the same not-yet-finished job returns the existing job id rather than queuing a duplicate.
      operationId: submit
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompileRequest"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
                  format: uuid
...
  /api/v1/conflicts/{copyId}/resolve:
    post:
      tags:
      - Conflicts
      summary: Resolve a conflict copy back into its original
      description: "The client renders the merge; no diff is computed server-side\
        \ (only the client's editor understands ProseMirror JSON). Resolving trashes\
        \ the copy — never deletes it, so a bad merge stays recoverable. Refuses a\
        \ stale baseVersion rather than forking again; the caller re-opens the updated\
        \ pair instead."
      operationId: resolve
      parameters:
      - name: copyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/comments/{commentId}/resolve:
    post:
      tags:
      - Comments
      operationId: resolve_1
      parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRequest"
        required: true
      responses:
        "204":
          description: No Content
...
  /api/v1/auth/refresh:
    post:
      tags:
      - Auth
      summary: Rotate a refresh token
      description: "Public. Refresh tokens rotate on every use and are single-use:\
        \ a leaked token works at most once, and the legitimate device's next refresh\
        \ failing is a detectable signal that it was stolen."
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SessionResponse"
  /api/v1/auth/password-reset:
    post:
      tags:
      - Account
      summary: Request a password reset email
      description: "Public. Always answers 202, whether or not the address is registered\
        \ — a different answer would make this an account-enumeration oracle. If mail\
        \ is not configured (spring.mail.host unset), the reset link is written to\
        \ the log instead of sent."
      operationId: requestReset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetRequest"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/auth/password-reset/confirm:
    post:
      tags:
      - Account
      summary: Complete a password reset
      description: "Public. Signs out every device holding a refresh token, reported\
        \ as devicesSignedOut."
      operationId: confirmReset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetConfirmRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
...
  /api/v1/account/deletion:
    get:
      tags:
      - Account
      operationId: deletionStatus
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeletionStatus"
    post:
      tags:
      - Account
      summary: Schedule account deletion
      description: "Requires the current password even though the caller is already\
        \ authenticated, so a borrowed unlocked laptop cannot destroy an account.\
        \ Deletion happens after a grace period (noveltea.account.deletion-grace),\
        \ cancellable via DELETE on this same path."
      operationId: requestDeletion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeletionStatus"
    delete:
      tags:
      - Account
      operationId: cancelDeletion
      responses:
        "204":
          description: No Content
...
  /api/v1/projects/{projectId}/conflicts:
    get:
      tags:
      - Conflicts
      operationId: list_3
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConflictSummary"
...
  /api/v1/projects/{projectId}/compile-jobs:
    get:
      tags:
      - Compile Jobs
      operationId: recent
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompileJob"
...
  /api/v1/conflicts/{copyId}:
    get:
      tags:
      - Conflicts
      operationId: detail
      parameters:
      - name: copyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictDetail"
  /api/v1/compile-jobs/{jobId}:
    get:
      tags:
      - Compile Jobs
      summary: Poll a compile job's status
      description: isTerminal on the returned job is what a client should loop on
        before calling .../download.
      operationId: status
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompileJob"
  /api/v1/compile-jobs/{jobId}/download:
    get:
      tags:
      - Compile Jobs
      summary: Download a finished compile job's file
      description: 410 (Gone) once a `download`-destination export has expired (noveltea.compile.download-ttl)
        or was never produced. A `server` destination export is kept indefinitely
        in the operator's mounted storage instead.
      operationId: download
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    JsonNode:
      type: object
...
    CompileRequest:
      type: object
      properties:
        format:
          type: string
        destination:
          type: string
        presetId:
          type: string
          format: uuid
        inlineConfig:
          $ref: "#/components/schemas/JsonNode"
...
    ResolveRequest:
      type: object
      properties:
        content:
          $ref: "#/components/schemas/JsonNode"
        baseVersion:
          type: integer
          format: int64
...
    SessionResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
    ResetRequest:
      type: object
      properties:
        email:
          type: string
    ResetConfirmRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
...
    DeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
    DeletionStatus:
      type: object
      properties:
        pending:
          type: boolean
        requestedAt:
          type: string
          format: date-time
        scheduledFor:
          type: string
          format: date-time
...
    ConflictSummary:
      type: object
      properties:
        copyId:
          type: string
          format: uuid
        originalId:
          type: string
          format: uuid
        originalTitle:
          type: string
        copyTitle:
          type: string
        forkedFromVersion:
          type: integer
          format: int64
        originalVersion:
          type: integer
          format: int64
        forkedAt:
          type: string
          format: date-time
    CompileJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        format:
          type: string
        destination:
          type: string
        status:
          type: string
        outputFilename:
          type: string
        outputBytes:
          type: integer
          format: int64
        wordCount:
          type: integer
          format: int32
        warnings:
          $ref: "#/components/schemas/JsonNode"
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        terminal:
          type: boolean
...
    ConflictDetail:
      type: object
      properties:
        copyId:
          type: string
          format: uuid
        originalId:
          type: string
          format: uuid
        originalTitle:
          type: string
        originalContent:
          $ref: "#/components/schemas/JsonNode"
        originalVersion:
          type: integer
          format: int64
        copyContent:
          $ref: "#/components/schemas/JsonNode"
        forkedFromVersion:
          type: integer
          format: int64
        forkedAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      description: "Access token from /auth/register, /auth/login, /auth/refresh or\
        \ /auth/pair. 15-minute lifetime by default (noveltea.auth.access-token-ttl)."
      scheme: bearer
      bearerFormat: JWT